Refactored code, extract hardcode from common module, to get it more flexible and reusable.
