+++ business-services/one/src/main/java/com/epam/javacc/microservices/servo/metrics/configuration/monitor/register/CounterMonitorRegister.java
@@ -1,14 +1,16 @@
-package com.epam.javacc.microservices.servo.metrics.configuration.monitor;
+package com.epam.javacc.microservices.servo.metrics.configuration.monitor.register;
 
-import com.epam.javacc.microservices.servo.metrics.common.configuration.MonitorMetricsRegistryBeanPostProcessor;
-import com.epam.javacc.microservices.servo.metrics.common.monitor.MonitorsKeeper;
 import com.epam.javacc.microservices.servo.metrics.common.monitor.MonitorRegister;
+import com.epam.javacc.microservices.servo.metrics.common.monitor.MonitorsKeeper;
 import com.netflix.servo.DefaultMonitorRegistry;
 import com.netflix.servo.monitor.BasicCounter;
 import com.netflix.servo.monitor.MonitorConfig;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
+
+import static com.epam.javacc.microservices.servo.metrics.configuration.metric.CommonMetricExtractor.REGISTRY_TAG_KEY;
+import static com.epam.javacc.microservices.servo.metrics.configuration.metric.CommonMetricExtractor.REGISTRY_TAG_VALUE;
 
 /**
  * Class for registering {@link BasicCounter} monitor metrics to the application.
