--- business-services/two-api/src/main/java/com/epam/javacc/microservices/servo/metrics/common/metric/MetricExtractor.java
@@ -1,79 +1,0 @@
-package com.epam.javacc.microservices.servo.metrics.common.metric;
-
-import com.netflix.servo.Metric;
-import com.netflix.servo.tag.TagList;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.function.Function;
-
-import static com.epam.javacc.microservices.servo.metrics.common.monitor.MonitorRegister.REGISTRY_TAG_KEY;
-import static com.epam.javacc.microservices.servo.metrics.common.monitor.MonitorRegister.REGISTRY_TAG_VALUE;
-import static java.util.stream.Collectors.groupingBy;
-import static java.util.stream.Collectors.toMap;
-
-@Component
-public class MetricExtractor {
-
-    private final double servoPollers;
-
-    MetricExtractor(@Value("${servo.pollers}") Double servoPollers) {
-        this.servoPollers = servoPollers;
-    }
-
-    @SuppressWarnings("unchecked")
-    public Map<String, Map<String,Object>> extractMetricsMapPerEndpoint(List<Metric> metrics) {
-        return metrics.stream()
-                .filter(this::isEndpointMonitorMetric)
-                .collect(groupingBy(m -> m.getConfig().getName()))
-                .entrySet().stream()
-                .map(e -> new Object[]{
-                        e.getKey(),
-                        e.getValue().stream()
-                                .map(this::mapMetric)
-                                .filter(Objects::nonNull)
-                                .collect(toMap(o1 -> (String) o1[0], (Function<Object[], Object>) o2 -> o2[1]))
-                }).collect(toMap(o -> (String) o[0], o -> (Map<String, Object>) o[1]));
-    }
-
-    private Object[] mapMetric(Metric m) {
-        TagList tags = m.getConfig().getTags();
-        if (isTotalCountMonitor(tags)) {
-            return new Object[]{
-                    "totalCount",
-                    m.getValue()
-            };
-        } else if (tags.containsKey("statistic") && tags.getValue("statistic").equals("totalTime")) {
-            //to obtain total method execution time within the polling interval
-            //we need to multiply totalTime by servo.pollers interval (in seconds)
-            return new Object[] {
-                    "totalExecutionTime",
-                    m.getNumberValue().doubleValue() * servoPollers + " ms."
-            };
-        } else if (tags.containsKey("statistic") && tags.getValue("statistic").equals("max")) {
-            return new Object[] {
-                    "maxExecutionTime",
-                    m.getValue() + " ms."
-            };
-        } else if (tags.containsKey("statistic") && tags.getValue("statistic").equals("count")) {
-            return new Object[] {
-                    "count",
-                    Math.round(m.getNumberValue().doubleValue() * servoPollers)
-            };
-        }
-        return null;
-    }
-
-    private static boolean isTotalCountMonitor(TagList tags) {
-        return tags.containsKey("type") && tags.getValue("type").equals("COUNTER");
-    }
-
-    private boolean isEndpointMonitorMetric(Metric metric) {
-        return  metric.getConfig().getTags().containsKey(REGISTRY_TAG_KEY) &&
-                metric.getConfig().getTags().getValue(REGISTRY_TAG_KEY).equals(REGISTRY_TAG_VALUE);
-    }
-
-}
